chore: move cash account files to a more sensible place
